refactoring  - extract interface from ProfilerConfig
