+++ test/src/main/java/com/navercorp/pinpoint/test/mock/MockTraceContext.java
@@ -16,6 +16,7 @@
 
 package com.navercorp.pinpoint.test.mock;
 
+import com.navercorp.pinpoint.bootstrap.config.DefaultProfilerConfig;
 import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
 import com.navercorp.pinpoint.bootstrap.context.AsyncTraceId;
 import com.navercorp.pinpoint.bootstrap.context.MethodDescriptor;
@@ -33,7 +34,7 @@ import com.navercorp.pinpoint.bootstrap.context.TraceType;
 public class MockTraceContext implements TraceContext {
 
     private Trace trace;
-    private final ProfilerConfig config = new ProfilerConfig();
+    private final ProfilerConfig config = new DefaultProfilerConfig();
 
     public void setTrace(Trace trace) {
         this.trace = trace;
